Add advanced MPlayer live ebuild.
This MPlayer ebuild is intended to be kept in sync with main tree
ebuild and send all possible patches there.

However it contains support for some out-of-the-main-tree packages
and rarely used features that are unlikely to be accepted by the
main repository, so it is placed here.

At this moment there are differences from portage:
- support for tivo (vstream-client);
- support for nemesi (but beware of broken nemesi manifests);
- use real binary codecs (in very rare cases when open codecs fail
  they may be usefull);
- restore support for svga video output and svga kernel helper.