Add openl2tp: http://openl2tp.sourceforge.net
This is the fastest openl2tp client and server available for Linux,
and it is the only one to utilize l2tp kernel module.

This ebuild was originally proposed at bug #195461.
I will add files since my first version in order to preserve change
history, so the following files are old historical data, but the
first files intended for version 1.7.